fix: ajuste no timeout e nas mensagens editadas
