+++ src/main/resources/application.properties
@@ -65,11 +65,11 @@ telegram.bot.username=${TELEGRAM_BOT_USERNAME}
 telegram.bot.token=${TELEGRAM_BOT_TOKEN}
 telegram.bot.polling.enabled=true
 # Timeouts para o cliente OkHttp (padrão: 30s)
-telegram.bot.client.connect-timeout=30
-telegram.bot.client.read-timeout=60
-telegram.bot.client.write-timeout=60
+telegram.bot.client.connect-timeout=60
+telegram.bot.client.read-timeout=120
+telegram.bot.client.write-timeout=120
 # Telegram Polling Timeouts
-telegram.bot.polling.timeout=60
+telegram.bot.polling.timeout=90
 telegram.bot.polling.allowed-updates=message,callback_query,edited_message
 
 telegram.message.limit=4000
